==> code/0a_check_instances_ara_m.ipynb <==
## Check to see if Melissa's gene pairs have feature information in the 2021 Cusack data

import pandas as pd

# Melissa's data
ara_m = pd.read_csv("../data/20240725_melissa_ara_data/interactions_fitness.txt", sep="\t")

# 2021 Cusack data
dat_2 = pd.read_csv("../data/2021_cusack_data/Dataset_2.txt", sep="\t") # inclusive genetic redundancy gene pairs
dat_4 = pd.read_csv("../data/2021_cusack_data/Dataset_4.txt", sep="\t") # protein kinase family gene pairs
features = pd.read_excel("../data/2021_cusack_data/Supplemental_tables_revision.xlsx",
                         sheet_name="Supplemental table 10", header=0)

print(ara_m.shape, dat_2.shape, dat_4.shape)

ara_m.head()

dat_2.head()

dat_4.head()

# split gene pair identifiers into two columns
dat_2_genes = dat_2.pair_ID.str.split('_', expand=True)
dat_4_genes = dat_4.pair_ID.str.split('_', expand=True)

# get gene pairs with feature information in dat_2 and dat_4
ara_m['MA'] = ara_m['MA'].str.upper() # convert gene names to uppercase
ara_m['MB'] = ara_m['MB'].str.upper()

ara_m_dat_2 = ara_m.loc[(ara_m.MA.isin(dat_2_genes[0]) | ara_m.MA.isin(dat_2_genes[1])) & \
    (ara_m.MB.isin(dat_2_genes[0]) | ara_m.MB.isin(dat_2_genes[1]))]

ara_m_dat_4 = ara_m.loc[(ara_m.MA.isin(dat_4_genes[0]) | ara_m.MA.isin(dat_4_genes[1])) & \
    (ara_m.MB.isin(dat_4_genes[0]) | ara_m.MB.isin(dat_4_genes[1]))]

ara_m_features = pd.concat([ara_m_dat_2, ara_m_dat_4], axis=0).drop_duplicates()
ara_m_features.to_csv(
  '../data/20240725_melissa_ara_data/interactions_fitness_gene_pairs_in_2021_cusack.txt',
  sep="\t", index=False)

ara_m.loc[~ara_m.Set.isin(ara_m_features.Set)].to_csv(
  '../data/20240725_melissa_ara_data/interactions_fitness_gene_pairs_NOT_in_2021_cusack.txt',
  sep="\t", index=False)

print(ara_m_features.shape)
